requirements.txt
langchain-openai
openai
langchain
python-dotenv
langchain_community
langchain_core
streamlit